app/worker/outreach-gcf-worker/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
supabase
langchain-core>=0.3.0
langchain-google-genai>=4.0.0
pydantic>=2.0.0